database/db/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "dart-backtest-db"
version = "0.1.0"
description = "Point-in-time-safe SQLite database for Korean equity backtesting"
requires-python = ">=3.10"
dependencies = [
    "pandas>=2.0.0",
    "openpyxl>=3.1.0",
    "xlrd>=2.0.1",
    "pyyaml>=6.0",
    "click>=8.1.0",
    "lxml>=4.9.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
